feat: 일정 삭제 로직 추가
